Fix template parameter for variable number of residual
The example had a const kNumObservations. Generally, when a truly
variable number of residuals exists, we have to use ceres::DYNAMIC and
pass the number of residuals to the cost function constructor.

Also, fix the issue with accumulating the residuals when Jet are
provided. The Jet is initialized with an invalid value, meaning it has
to be assigned and not serve as an initial accumulation value.

Signed-off-by: Francis Giraldeau <francis.giraldeau@nrc-cnrc.gc.ca>
Change-Id: I02d6fd785911db3ea54f19fba82e5dcc665c85c1